Adjust binary file inserts after schema changes
We have changed the way attachments are linked to a version.

The table `ldml_version_attachment` was replaced with `ldml_version_norm_xml` and each `ldml_version` now has a many-to-many relationship to `norm_xml`. 

Furthermore attachments are now directly linked to its corresponding `norm_xml` via a many-to-one relationship (from the attachment side).